fix(prometheus.operator): Address scrape class review feedback
- Move the duplicated scrape_class docs into docs/sources/shared so all
  four component pages import them instead of repeating the markdown.
- Fold the scrape class tests into the per-generator test files and drop
  config_gen_scrapeclass_more_test.go.
- Guard the Probe BearerTokenSecret pointer in epHasAuth.
- Add an end-to-end scrape class test alongside TestServiceMonitorEndToEnd,
  extracting the component setup into a shared helper.
