+++ internal/component/prometheus/operator/configgen/config_gen_podmonitor_test.go
@@ -590,3 +590,26 @@ func TestGeneratePodMonitorConfig(t *testing.T) {
 		})
 	}
 }
+
+func TestGeneratePodMonitorConfigScrapeClass(t *testing.T) {
+	cg := scrapeClassTestGenerator()
+	m := &promopv1.PodMonitor{ObjectMeta: metav1.ObjectMeta{Namespace: "ns", Name: "pm"}}
+	ep := promopv1.PodMetricsEndpoint{
+		RelabelConfigs:       []promopv1.RelabelConfig{{TargetLabel: "from_endpoint", Replacement: ptr.To("yes")}},
+		MetricRelabelConfigs: []promopv1.RelabelConfig{{TargetLabel: "metric_from_endpoint", Replacement: ptr.To("yes")}},
+	}
+
+	cfg, err := cg.GeneratePodMonitorConfig(m, ep, 0)
+	require.NoError(t, err)
+
+	assert.True(t, cfg.HTTPClientConfig.TLSConfig.InsecureSkipVerify)
+	require.NotNil(t, cfg.HTTPClientConfig.Authorization)
+	assert.Equal(t, "class-token", string(cfg.HTTPClientConfig.Authorization.Credentials))
+
+	sd, ok := cfg.ServiceDiscoveryConfigs[0].(*promk8s.SDConfig)
+	require.True(t, ok)
+	assert.True(t, sd.AttachMetadata.Node)
+
+	assert.Less(t, indexOfTargetLabel(cfg.RelabelConfigs, "from_class"), indexOfTargetLabel(cfg.RelabelConfigs, "from_endpoint"))
+	assert.Greater(t, indexOfTargetLabel(cfg.MetricRelabelConfigs, "metric_from_class"), indexOfTargetLabel(cfg.MetricRelabelConfigs, "metric_from_endpoint"))
+}
