+++ internal/component/prometheus/operator/configgen/config_gen_servicemonitor_test.go
@@ -669,3 +669,64 @@ func TestGenerateServiceMonitorConfig(t *testing.T) {
 		})
 	}
 }
+
+func TestGenerateServiceMonitorConfigScrapeClassDefaultApplied(t *testing.T) {
+	cg := scrapeClassTestGenerator()
+	m := &promopv1.ServiceMonitor{ObjectMeta: metav1.ObjectMeta{Namespace: "ns", Name: "sm"}}
+	ep := promopv1.Endpoint{
+		RelabelConfigs:       []promopv1.RelabelConfig{{TargetLabel: "from_endpoint", Replacement: ptr.To("yes")}},
+		MetricRelabelConfigs: []promopv1.RelabelConfig{{TargetLabel: "metric_from_endpoint", Replacement: ptr.To("yes")}},
+	}
+
+	cfg, err := cg.GenerateServiceMonitorConfig(m, ep, 0, promk8s.RoleEndpoint)
+	require.NoError(t, err)
+
+	assert.True(t, cfg.HTTPClientConfig.TLSConfig.InsecureSkipVerify)
+	require.NotNil(t, cfg.HTTPClientConfig.Authorization)
+	assert.Equal(t, "class-token", string(cfg.HTTPClientConfig.Authorization.Credentials))
+
+	sd, ok := cfg.ServiceDiscoveryConfigs[0].(*promk8s.SDConfig)
+	require.True(t, ok)
+	assert.True(t, sd.AttachMetadata.Node)
+
+	classIdx := indexOfTargetLabel(cfg.RelabelConfigs, "from_class")
+	epIdx := indexOfTargetLabel(cfg.RelabelConfigs, "from_endpoint")
+	require.NotEqual(t, -1, classIdx)
+	require.NotEqual(t, -1, epIdx)
+	assert.Less(t, classIdx, epIdx, "class relabelings should be prepended before endpoint relabelings")
+
+	classMetricIdx := indexOfTargetLabel(cfg.MetricRelabelConfigs, "metric_from_class")
+	epMetricIdx := indexOfTargetLabel(cfg.MetricRelabelConfigs, "metric_from_endpoint")
+	require.NotEqual(t, -1, classMetricIdx)
+	require.NotEqual(t, -1, epMetricIdx)
+	assert.Greater(t, classMetricIdx, epMetricIdx, "class metric relabelings should be appended after endpoint metric relabelings")
+}
+
+func TestGenerateServiceMonitorConfigScrapeClassEndpointOverridesTLS(t *testing.T) {
+	cg := scrapeClassTestGenerator()
+	m := &promopv1.ServiceMonitor{ObjectMeta: metav1.ObjectMeta{Namespace: "ns", Name: "sm"}}
+	// The endpoint sets its own (empty) TLS config, so the class TLS must not apply.
+	ep := promopv1.Endpoint{
+		HTTPConfigWithProxyAndTLSFiles: promopv1.HTTPConfigWithProxyAndTLSFiles{
+			HTTPConfigWithTLSFiles: promopv1.HTTPConfigWithTLSFiles{
+				TLSConfig: &promopv1.TLSConfig{},
+			},
+		},
+	}
+
+	cfg, err := cg.GenerateServiceMonitorConfig(m, ep, 0, promk8s.RoleEndpoint)
+	require.NoError(t, err)
+
+	assert.False(t, cfg.HTTPClientConfig.TLSConfig.InsecureSkipVerify, "endpoint TLS should take precedence over the class")
+}
+
+func TestGenerateServiceMonitorConfigScrapeClassNotDefined(t *testing.T) {
+	cg := scrapeClassTestGenerator()
+	m := &promopv1.ServiceMonitor{
+		ObjectMeta: metav1.ObjectMeta{Namespace: "ns", Name: "sm"},
+		Spec:       promopv1.ServiceMonitorSpec{ScrapeClassName: ptr.To("does-not-exist")},
+	}
+
+	_, err := cg.GenerateServiceMonitorConfig(m, promopv1.Endpoint{}, 0, promk8s.RoleEndpoint)
+	require.Error(t, err)
+}
